Refactor almost finished, ready for debugging. null pointer exception in teleopinit. Isaac
