+++ src/org/usfirst/frc/team5985/robot/DriverStation.java
@@ -1,26 +1,25 @@
 package org.usfirst.frc.team5985.robot;
 
-import edu.wpi.first.wpilibj.Joystick;
-import edu.wpi.first.wpilibj.Joystick.RumbleType;
-import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 import edu.wpi.first.wpilibj.CameraServer;
+import edu.wpi.first.wpilibj.Joystick;
+import edu.wpi.first.wpilibj.smartdashboard.SmartDashboard;
 
 public class DriverStation {
 
 	Joystick stick;
 	Joystick xbox;
 	
-	public void teleopInit(Robot robot) {
+	public void driverInit(CameraServer camera1) {
 		
 		stick = new Joystick(0);
         xbox =	 new Joystick(1);
         
-        robot.camera1 = CameraServer.getInstance();
-    	robot.camera1.setQuality(50);
-    	robot.camera1.startAutomaticCapture("cam0");
+        camera1 = CameraServer.getInstance();
+    	camera1.setQuality(50);
+    	camera1.startAutomaticCapture("cam0");
 		
 	}
-	public void teleopPeriodic() {
+	public void driverPeriodic() {
 		
 		//Feeds stick + xbox info to Smart Dashboard
 		
@@ -29,6 +28,7 @@ public class DriverStation {
     	SmartDashboard.putBoolean("Stick Button 1 (Intake In)", stick.getRawButton(1));
     	SmartDashboard.putBoolean("Stick Button 2 (Intake Out)", stick.getRawButton(2));
     	SmartDashboard.putNumber("Xbox Y Value (Arm)", xbox.getRawAxis(1));
+
